+++ Player.java
@@ -3,19 +3,19 @@ import java.io.Serializable;
 public class Player implements Serializable{
     private static final long serialVersionUID = 13412431243L;
 
-    int x;
-    int y;
+    double x;
+    double y;
     double rotation;
     int[] guns = new int[2];
     String name;
 
-    public Player(int x, int y, double rotation) {
+    public Player(double x, double y, double rotation) {
         this.x = x;
         this.y = y;
         this.rotation = rotation;
     }
 
-    public void setLocation (int[] location) {
+    public void setLocation (double[] location) {
         this.x = location[0];
         this.y = location[1];
     }
